removed unnecessary classes
