+++ data-processing-lib/src/data_processing/runtime/pure_python/__init__.py
@@ -1,4 +1,3 @@
-from data_processing.runtime.pure_python.python_launcher_configuration import PythonLauncherConfiguration
 from data_processing.runtime.pure_python.transform_table_processor import TransformTableProcessor
 from data_processing.runtime.pure_python.transform_orchestrator import orchestrate
 from data_processing.runtime.pure_python.transform_launcher import PythonTransformLauncher
