+++ data-processing-lib/src/data_processing/runtime/ray/__init__.py
@@ -1,8 +1,8 @@
 from data_processing.runtime.ray.ray_utils import RayUtils
 from data_processing.runtime.ray.transform_statistics import TransformStatisticsRay
-from data_processing.runtime.ray.transform_table_processor import TransformTableProcessorRay
 from data_processing.runtime.ray.transform_runtime import DefaultTableTransformRuntimeRay
-from data_processing.runtime.ray.transform_launch_configuration import RayLauncherConfiguration
+from data_processing.runtime.ray.transform_configuration import RayTransformConfiguration
+from data_processing.runtime.ray.transform_table_processor import TransformTableProcessorRay
 from data_processing.runtime.ray.transform_orchestrator_configuration import TransformOrchestratorConfiguration
 from data_processing.runtime.ray.transform_orchestrator import orchestrate
 from data_processing.runtime.ray.transform_launcher import RayTransformLauncher
