+++ data-processing-lib/src/data_processing/runtime/ray/transform_launcher.py
@@ -17,8 +17,6 @@ import time
 import ray
 from data_processing.data_access import DataAccessFactory, DataAccessFactoryBase
 from data_processing.runtime.ray import (
-    DefaultTableTransformRuntimeRay,
-    RayLauncherConfiguration,
     TransformOrchestratorConfiguration,
     orchestrate,
 )
@@ -26,7 +24,6 @@ from data_processing.runtime.ray.transform_configuration import (
     RayTransformConfiguration,
 )
 from data_processing.runtime.transform_launcher import AbstractTransformLauncher
-from data_processing.transform import TransformConfiguration
 from data_processing.utils import get_logger, str2bool
 
 
@@ -41,16 +38,15 @@ class RayTransformLauncher(AbstractTransformLauncher):
     def __init__(
         self,
         transform_config: RayTransformConfiguration,
-        # runtime_class: type[DefaultTableTransformRuntimeRay]=DefaultTableTransformRuntimeRay,
         data_access_factory: DataAccessFactoryBase = DataAccessFactory(),
     ):
         """
         Creates driver
-        :param transform_runtime_config: transform runtime factory
+        :param transform_config: transform runtime factory
         :param data_access_factory: the factory to create DataAccess instances.
         """
         super().__init__(transform_config, data_access_factory)
-        self.transform_runtime_config = RayLauncherConfiguration(transform_config, transform_config.runtime_class)
+        self.transform_runtime_config = transform_config
         self.ray_orchestrator = TransformOrchestratorConfiguration(name=self.name)
 
     def __get_parameters(self) -> bool:
