+++ data-processing-lib/test/data_processing_tests/launch/pure_python/launcher_test.py
@@ -13,10 +13,7 @@
 import os
 import sys
 
-from data_processing.runtime.pure_python import (
-    PythonLauncherConfiguration,
-    PythonTransformLauncher,
-)
+from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.transform import AbstractTableTransform, TransformConfiguration
 from data_processing.utils import ParamsUtils
