+++ data-processing-lib/test/data_processing_tests/launch/ray/launcher_test.py
@@ -14,14 +14,9 @@ import os
 import sys
 
 from data_processing.runtime.ray import RayTransformLauncher
-from data_processing.runtime.ray.transform_configuration import (
-    RayTransformConfiguration,
-)
-from data_processing.test_support.transform import NOOPTransformConfiguration
 from data_processing.test_support.transform.noop_transform import (
     NOOPRayTransformConfiguration,
 )
-from data_processing.transform import AbstractTableTransform, TransformConfiguration
 from data_processing.utils import ParamsUtils
