+++ transforms/code/malware/src/malware_transform.py
@@ -18,10 +18,6 @@ from typing import Any
 
 import clamd
 import pyarrow as pa
-from data_processing.runtime.pure_python import (
-    PythonLauncherConfiguration,
-    PythonTransformLauncher,
-)
 from data_processing.runtime.ray import RayTransformLauncher
 from data_processing.runtime.ray.transform_configuration import (
     RayTransformConfiguration,
